added missing features

+++ app/src/main/java/com/example/cryptocurrencytracker/model/dto/CoinHistoryDto.kt
@@ -10,8 +10,8 @@ data class CoinHistoryDto(
     val time: Long,
     @SerializedName("open")
     @Expose
-    val open: Float,
+    val open: Double,
     @SerializedName("close")
     @Expose
-    val close: Float
+    val close: Double
 )
